requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==3.2.20
djangorestframework==3.14.0
httpie==3.2.2
idna==3.6
importlib-metadata==6.7.0
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
# numpy==1.21.6
# pandas==1.3.5
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
zipp==3.15.0
